Separate out classes formerly testing StringTemplate3StatementLocator into separate classes now that `<foo>` interpolation is in core.
